pyproject.toml
[tool.poetry]
name = "EffiCare"
version = "0.1.0"
description = "Learning representations for ICU"

license = "MIT"

authors = [
    "Oğuz Şerbetci <oguz.serbetci@gmail.com>"
]

readme = 'README.md'

[tool.poetry.dependencies]
python = "~3.7"
torch = "1.4.0"
torchtext = "0.4.0"
tensorboard = "2.0.0"
gitpython = "3.1.1"
pytorch-ignite = "0.4.1"
scikit-learn = "0.22.2"
seaborn = "0.10.0"
wandb = "0.8.34"
colorlog = "4.1.0"
matplotlib = "3.2.1"
plotly = "4.7.1"
captum = "0.2.0"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
ipykernel = "^5.2.1"
mypy = "^0.770"
pyflakes = "^2.2.0"

[flake8]
ignore = "E501"
